+++ src/automation/HISTORY.rst
@@ -5,7 +5,10 @@ Release History
 
 0.2.0
 ++++++
-* `az automation`: Add new command group
+* `az automation schedule`: Add new command group to support managing schedule
+* `az automation software-update-configuration`: Add new command group support managing software-update-configuration
+* `az automation software-update-configuration runs`: Add new command group support managing software-update-configuration runs
+* `az automation software-update-configuration machine-runs`: Add new command group support managing software-update-configuration machine-runs
 
 0.1.0
 ++++++